Create Number Guessing Game
Now this is just a simple game where the user needs to guess the number the system has randomly generated. There 3 rounds where the user has 5 attempts each to guess the number, and based on the users' input the computer lets them know whether that number is greater than or less than the actual number. At the end of these 3 rounds the answer is revealed if the user isn't able to get the correct answer.